requirements.txt
pandas
matplotlib
plotly
SQLAlchemy
openpyxl
python-dateutil
psycopg2-binary
python-dotenv